backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-dateutil==2.8.2
pytz==2023.3
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
websockets==12.0
redis[asyncio]==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scipy==1.11.4
aioredis==2.0.1
ccxt==4.0.77
